Bug da Home duplic, Orientacao e del de img

+++ lib/models/section/section.dart
@@ -65,7 +65,8 @@ class Section extends ChangeNotifier {
     }
 
     for (final original in originalItems) {
-      if (!items.contains(original)) {
+      if (!items.contains(original) &&
+          (original.image as String).contains('firebase')) {
         try {
           final ref = await storage.ref(original.image as String);
           await ref.delete();
@@ -84,11 +85,13 @@ class Section extends ChangeNotifier {
   Future<void> delete() async {
     await firestoreRef.delete();
     for (final item in items) {
-      try {
-        final ref = await storage.ref(item.image as String);
-        await ref.delete();
-        // ignore: empty_catches
-      } catch (e) {}
+      if ((item.image as String).contains('firebase')) {
+        try {
+          final ref = await storage.ref(item.image as String);
+          await ref.delete();
+          // ignore: empty_catches
+        } catch (e) {}
+      }
     }
   }
